fix ESLint and build error on addPlantPage
